Fix memory leak and unhandled memory fault in `schema-from-object`
The memory leak was a result of not freeing g-lists which we created,
and the unhandled memory fault was from the garbage collector clearing
the data elements in a g-list before `make-arrow-schema-new` could
enumerate the list and pull out its elements.

This commit also introduces a test suite.

Fixes #2
Fixes #3